Add groupName to links and refactor out query if-else chain in link controller * Remove unnecessary type conversion from link generator

+++ spring18/src/main/java/com/sylconnexity/spring18/util/ClickGenerator.java
@@ -39,7 +39,7 @@ public class ClickGenerator extends RandomGenerator {
         } else {
             orderAmount = 0.0;
             orderNumber = null;
-            unitsOrdered = (long) 0;
+            unitsOrdered = 0L;
         }
 
         String randomDate = generateRandomDate();
